Refactor command handling to use 32-bit integers for opcodes and data types
- Updated all instances of push_command to accept uint32_t instead of uint8_t for opcodes across various modules including audio, canvas, dom, input, storage, system, webgl, websocket, and webcc.
- Changed data handling in push_data to support uint32_t, int32_t, and float types with appropriate serialization methods.
- Adjusted command buffer implementation to ensure proper alignment and padding for 32-bit data types.
- Modified JavaScript bindings to read and handle 32-bit integers and floats correctly, ensuring compatibility with the updated command structure.
- Enhanced error handling for out-of-bounds access in JavaScript when reading from the command buffer.

+++ include/webcc/audio.h
@@ -20,25 +20,25 @@ namespace webcc::audio {
     }
 
     inline void play(int32_t handle){
-        push_command((uint8_t)OP_PLAY);
+        push_command((uint32_t)OP_PLAY);
         push_data<int32_t>(handle);
     }
 
     inline void pause(int32_t handle){
-        push_command((uint8_t)OP_PAUSE);
+        push_command((uint32_t)OP_PAUSE);
         push_data<int32_t>(handle);
     }
 
     inline void set_volume(int32_t handle, float vol){
-        push_command((uint8_t)OP_SET_VOLUME);
+        push_command((uint32_t)OP_SET_VOLUME);
         push_data<int32_t>(handle);
         push_data<float>(vol);
     }
 
     inline void set_loop(int32_t handle, uint8_t loop){
-        push_command((uint8_t)OP_SET_LOOP);
+        push_command((uint32_t)OP_SET_LOOP);
         push_data<int32_t>(handle);
-        push_data<uint8_t>(loop);
+        push_data<uint32_t>((uint32_t)loop);
     }
 
     extern "C" float webcc_audio_get_current_time(int32_t handle);
