+++ include/webcc/input.h
@@ -27,21 +27,21 @@ namespace webcc::input {
     };
 
     inline void init_keyboard(){
-        push_command((uint8_t)OP_INIT_KEYBOARD);
+        push_command((uint32_t)OP_INIT_KEYBOARD);
     }
 
     inline void init_mouse(int32_t handle){
-        push_command((uint8_t)OP_INIT_MOUSE);
+        push_command((uint32_t)OP_INIT_MOUSE);
         push_data<int32_t>(handle);
     }
 
     inline void request_pointer_lock(int32_t handle){
-        push_command((uint8_t)OP_REQUEST_POINTER_LOCK);
+        push_command((uint32_t)OP_REQUEST_POINTER_LOCK);
         push_data<int32_t>(handle);
     }
 
     inline void exit_pointer_lock(){
-        push_command((uint8_t)OP_EXIT_POINTER_LOCK);
+        push_command((uint32_t)OP_EXIT_POINTER_LOCK);
     }
 
 } // namespace webcc::input
