+++ include/webcc/webgl.h
@@ -32,7 +32,7 @@ namespace webcc::webgl {
     }
 
     inline void viewport(int32_t ctx_handle, int32_t x, int32_t y, int32_t width, int32_t height){
-        push_command((uint8_t)OP_VIEWPORT);
+        push_command((uint32_t)OP_VIEWPORT);
         push_data<int32_t>(ctx_handle);
         push_data<int32_t>(x);
         push_data<int32_t>(y);
@@ -41,7 +41,7 @@ namespace webcc::webgl {
     }
 
     inline void clear_color(int32_t ctx_handle, float r, float g, float b, float a){
-        push_command((uint8_t)OP_CLEAR_COLOR);
+        push_command((uint32_t)OP_CLEAR_COLOR);
         push_data<int32_t>(ctx_handle);
         push_data<float>(r);
         push_data<float>(g);
@@ -50,7 +50,7 @@ namespace webcc::webgl {
     }
 
     inline void clear(int32_t ctx_handle, uint32_t mask){
-        push_command((uint8_t)OP_CLEAR);
+        push_command((uint32_t)OP_CLEAR);
         push_data<int32_t>(ctx_handle);
         push_data<uint32_t>(mask);
     }
@@ -68,20 +68,20 @@ namespace webcc::webgl {
     }
 
     inline void attach_shader(int32_t ctx_handle, int32_t prog_handle, int32_t shader_handle){
-        push_command((uint8_t)OP_ATTACH_SHADER);
+        push_command((uint32_t)OP_ATTACH_SHADER);
         push_data<int32_t>(ctx_handle);
         push_data<int32_t>(prog_handle);
         push_data<int32_t>(shader_handle);
     }
 
     inline void link_program(int32_t ctx_handle, int32_t prog_handle){
-        push_command((uint8_t)OP_LINK_PROGRAM);
+        push_command((uint32_t)OP_LINK_PROGRAM);
         push_data<int32_t>(ctx_handle);
         push_data<int32_t>(prog_handle);
     }
 
     inline void bind_attrib_location(int32_t ctx_handle, int32_t prog_handle, uint32_t index, const char* name){
-        push_command((uint8_t)OP_BIND_ATTRIB_LOCATION);
+        push_command((uint32_t)OP_BIND_ATTRIB_LOCATION);
         push_data<int32_t>(ctx_handle);
         push_data<int32_t>(prog_handle);
         push_data<uint32_t>(index);
@@ -89,7 +89,7 @@ namespace webcc::webgl {
     }
 
     inline void use_program(int32_t ctx_handle, int32_t prog_handle){
-        push_command((uint8_t)OP_USE_PROGRAM);
+        push_command((uint32_t)OP_USE_PROGRAM);
         push_data<int32_t>(ctx_handle);
         push_data<int32_t>(prog_handle);
     }
@@ -101,14 +101,14 @@ namespace webcc::webgl {
     }
 
     inline void bind_buffer(int32_t ctx_handle, uint32_t target, int32_t buf_handle){
-        push_command((uint8_t)OP_BIND_BUFFER);
+        push_command((uint32_t)OP_BIND_BUFFER);
         push_data<int32_t>(ctx_handle);
         push_data<uint32_t>(target);
         push_data<int32_t>(buf_handle);
     }
 
     inline void buffer_data(int32_t ctx_handle, uint32_t target, uint32_t data_ptr, uint32_t data_len, uint32_t usage){
-        push_command((uint8_t)OP_BUFFER_DATA);
+        push_command((uint32_t)OP_BUFFER_DATA);
         push_data<int32_t>(ctx_handle);
         push_data<uint32_t>(target);
         push_data<uint32_t>(data_ptr);
@@ -117,13 +117,13 @@ namespace webcc::webgl {
     }
 
     inline void enable_vertex_attrib_array(int32_t ctx_handle, uint32_t index){
-        push_command((uint8_t)OP_ENABLE_VERTEX_ATTRIB_ARRAY);
+        push_command((uint32_t)OP_ENABLE_VERTEX_ATTRIB_ARRAY);
         push_data<int32_t>(ctx_handle);
         push_data<uint32_t>(index);
     }
 
     inline void enable(int32_t ctx_handle, uint32_t cap){
-        push_command((uint8_t)OP_ENABLE);
+        push_command((uint32_t)OP_ENABLE);
         push_data<int32_t>(ctx_handle);
         push_data<uint32_t>(cap);
     }
@@ -135,25 +135,25 @@ namespace webcc::webgl {
     }
 
     inline void uniform_1f(int32_t ctx_handle, int32_t loc_handle, float val){
-        push_command((uint8_t)OP_UNIFORM_1F);
+        push_command((uint32_t)OP_UNIFORM_1F);
         push_data<int32_t>(ctx_handle);
         push_data<int32_t>(loc_handle);
         push_data<float>(val);
     }
 
     inline void vertex_attrib_pointer(int32_t ctx_handle, uint32_t index, int32_t size, uint32_t type, uint8_t normalized, int32_t stride, int32_t offset){
-        push_command((uint8_t)OP_VERTEX_ATTRIB_POINTER);
+        push_command((uint32_t)OP_VERTEX_ATTRIB_POINTER);
         push_data<int32_t>(ctx_handle);
         push_data<uint32_t>(index);
         push_data<int32_t>(size);
         push_data<uint32_t>(type);
-        push_data<uint8_t>(normalized);
+        push_data<uint32_t>((uint32_t)normalized);
         push_data<int32_t>(stride);
         push_data<int32_t>(offset);
     }
 
     inline void draw_arrays(int32_t ctx_handle, uint32_t mode, int32_t first, int32_t count){
-        push_command((uint8_t)OP_DRAW_ARRAYS);
+        push_command((uint32_t)OP_DRAW_ARRAYS);
         push_data<int32_t>(ctx_handle);
         push_data<uint32_t>(mode);
         push_data<int32_t>(first);
